deploy: rename deployment controller to deployer controller
